+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Feature: Add new `Progress` React component.
 - Feature: Add 'PhoneIllustration' React component.
+- Fix: Propagate props from `CallToActionPopover` to `Popover` components.
+- Fix: Add key to buttons rendering in `CallToActionPopover`.
 
 ## [5.9.0] - 2017-02-06